# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Test Project",
        "description": "",
        "termsOfService": "",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Laravel test project server."
        }
    ],
    "paths": {
...
        "/api/reservations/{reservationId}": {
            "get": {
                "tags": [
                    "reservations.general"
                ],
                "summary": "Get details of a reservation.",
                "description": "",
                "operationId": "reservations.show",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/reservations/{reservationId}/room": {
            "get": {
                "tags": [
                    "reservations.general"
                ],
                "summary": "Get details of related room.",
                "description": "",
                "operationId": "reservations.rooms.show",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/reservations/status": {
            "patch": {
                "tags": [
                    "reservations.status"
                ],
                "summary": "Sets reservation status",
                "description": "",
                "operationId": "reservations.status.patch",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationPATCH"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationGET"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/rooms/{roomId}": {
            "get": {
                "tags": [
                    "rooms.general"
                ],
                "summary": "Get details of a room.",
                "description": "",
                "operationId": "rooms.show",
                "parameters": [
                    {
                        "name": "roomId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        },
        "/api/rooms": {
            "post": {
                "tags": [
                    "rooms.general"
                ],
                "summary": "Create new room.",
                "description": "",
                "operationId": "rooms.store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/vnd.api+json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomPOST"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoomGET"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "deprecated": false
            }
        }
    },
    "components": {
        "schemas": {
...
            "ReservationGETAttributes": {
                "required": [
                    "from",
                    "to",
                    "adults_amount",
                    "children_amount",
                    "infants_amount",
                    "first_name",
                    "last_name",
                    "email",
                    "phone",
                    "notes",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "from": {
                        "type": "string",
                        "format": "Y-m-d",
                        "example": "2020-01-01"
                    },
                    "to": {
                        "type": "string",
                        "format": "Y-m-d",
                        "example": "2020-01-03"
                    },
                    "adults_amount": {
                        "type": "number",
                        "format": "int32",
                        "maximum": 4,
                        "minimum": 0,
                        "example": "2"
                    },
                    "children_amount": {
                        "type": "number",
                        "format": "int32",
                        "maximum": 4,
                        "minimum": 0,
                        "example": "0"
                    },
                    "infants_amount": {
                        "type": "number",
                        "format": "int32",
                        "maximum": 4,
                        "minimum": 0,
                        "example": "0"
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3,
                        "example": "john.doe@example.com"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 65535,
                        "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry."
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "cancelled",
                            "completed",
                            "overdue"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "Y-m-d H:i:s",
                        "example": "2020-01-01 11:55:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "Y-m-d H:i:s",
                        "example": "2020-01-01 11:55:00"
                    }
                },
                "type": "object"
            },
            "ReservationGETDataItemLinks": {
                "required": [
                    "self"
                ],
                "properties": {
                    "self": {
                        "type": "string",
                        "example": "http://localhost/api/reservations/1"
                    }
                },
                "type": "object"
            },
            "ReservationGETRelationships": {
                "required": [
                    "room"
                ],
                "properties": {
                    "room": {
                        "description": "",
                        "type": "object",
                        "$ref": "#/components/schemas/RoomRelatedSingleOutput"
                    }
                },
                "type": "object"
            },
            "ReservationGETDataItem": {
                "required": [
                    "id",
                    "type",
                    "attributes",
                    "relationships",
                    "links"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "reservations"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationGETAttributes"
                    },
                    "relationships": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationGETRelationships"
                    },
                    "links": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationGETDataItemLinks"
                    }
                },
                "type": "object"
            },
            "ReservationGET": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationGETDataItem"
                    }
                },
                "type": "object"
            },
            "ReservationPatchAttributes": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "cancelled",
                            "completed",
                            "overdue"
                        ],
                        "example": "pending"
                    }
                },
                "type": "object"
            },
            "ReservationPATCHDataItem": {
                "required": [
                    "type",
                    "id",
                    "attributes"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "reservations"
                    },
                    "id": {
                        "type": "int",
                        "example": "1"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationPatchAttributes"
                    }
                },
                "type": "object"
            },
            "ReservationPATCH": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/ReservationPATCHDataItem"
                    }
                },
                "type": "object"
            },
            "RoomGETAttributes": {
                "required": [
                    "type",
                    "number",
                    "floor",
                    "price_default",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "superior",
                            "executive",
                            "suite"
                        ]
                    },
                    "number": {
                        "type": "string",
                        "maxLength": 10,
                        "minLength": 3,
                        "example": "201"
                    },
                    "floor": {
                        "type": "string",
                        "enum": [
                            "2",
                            "3",
                            "4"
                        ]
                    },
                    "price_default": {
                        "type": "number",
                        "format": "float",
                        "example": "200.00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "Y-m-d H:i:s",
                        "example": "2020-01-01 11:55:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "Y-m-d H:i:s",
                        "example": "2020-01-01 11:55:00"
                    }
                },
                "type": "object"
            },
            "RoomGETDataItemLinks": {
                "required": [
                    "self"
                ],
                "properties": {
                    "self": {
                        "type": "string",
                        "example": "http://localhost/api/rooms/1"
                    }
                },
                "type": "object"
            },
            "RoomGETDataItem": {
                "required": [
                    "id",
                    "type",
                    "attributes",
                    "links"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "rooms"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomGETAttributes"
                    },
                    "links": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomGETDataItemLinks"
                    }
                },
                "type": "object"
            },
            "RoomGET": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomGETDataItem"
                    }
                },
                "type": "object"
            },
            "RoomRelatedDataObject": {
                "required": [
                    "id",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "type": {
                        "type": "string",
                        "example": "rooms"
                    }
                },
                "type": "object"
            },
            "RoomRelatedLinkObject": {
                "required": [
                    "related"
                ],
                "properties": {
                    "related": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RoomRelatedSingleOutput": {
                "required": [
                    "data"
                ],
                "properties": {
                    "links": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomRelatedLinkObject"
                    },
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomRelatedDataObject"
                    }
                },
                "type": "object"
            },
            "RoomPOSTAttributes": {
                "required": [
                    "type",
                    "number",
                    "floor",
                    "price_default"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "superior",
                            "executive",
                            "suite"
                        ]
                    },
                    "number": {
                        "type": "string",
                        "maxLength": 10,
                        "minLength": 3,
                        "example": "201"
                    },
                    "floor": {
                        "type": "string",
                        "enum": [
                            "2",
                            "3",
                            "4"
                        ]
                    },
                    "price_default": {
                        "type": "number",
                        "format": "float",
                        "example": "200.00"
                    }
                },
                "type": "object"
            },
            "RoomPOSTDataItem": {
                "required": [
                    "type",
                    "attributes"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "rooms"
                    },
                    "attributes": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomPOSTAttributes"
                    }
                },
                "type": "object"
            },
            "RoomPOST": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/RoomPOSTDataItem"
                    }
                },
                "type": "object"
            }
        }
    },
...
}